refactor: move ChangeOrthography view to morphodict app
