+++ CreeDictionary/CreeDictionary/urls.py
@@ -1,8 +1,6 @@
 """
 Definition of urls for CreeDictionary.
 """
-import API.views as api_views
-from CreeDictionary import views
 from django.conf import settings
 from django.conf.urls import url
 from django.conf.urls.static import static
@@ -10,6 +8,9 @@ from django.contrib import admin
 from django.contrib.staticfiles.urls import staticfiles_urlpatterns
 from django.urls import include, path
 from django_js_reverse.views import urls_js
+
+import API.views as api_views
+from CreeDictionary import views
 
 # 2019/May/21 Matt Yan:
 
@@ -63,11 +64,7 @@ _urlpatterns = [
         "cree-dictionary-word-translation-api",
     ),
     ("admin/", admin.site.urls, "admin"),
-    (
-        "change-orthography",
-        views.ChangeOrthography.as_view(),
-        "cree-dictionary-change-orthography",
-    ),
+    ("", include("morphodict.urls"), "cree-dictionary-change-orthography",),
 ]
 
 # XXX: ugly hack to make this work on a local instance and on Sapir
