+++ CreeDictionary/CreeDictionary/views.py
@@ -2,12 +2,10 @@ from http import HTTPStatus
 
 from django.http import HttpResponse, HttpResponseBadRequest, HttpResponseNotFound
 from django.shortcuts import redirect, render
-from django.views import View
 from django.views.decorators.http import require_GET
 
 from API.models import Wordform
 from CreeDictionary.forms import WordSearchForm
-from morphodict.orthography import ORTHOGRAPHY
 from utils import ParadigmSize
 
 from .utils import url_for_query
@@ -168,35 +166,6 @@ def contact_us(request):  # pragma: no cover
     Contact us page.
     """
     return render(request, "CreeDictionary/contact-us.html")
-
-
-class ChangeOrthography(View):
-    """
-    Sets the orth= cookie, which affects the default rendered orthography.
-
-        > POST /change-orthography HTTP/1.1
-        > Cookie: orth=Latn
-        >
-        > orth=Cans
-
-        < HTTP/1.1 204 No Content
-        < Set-Cookie: orth=Cans
-
-    Supports only POST requests for now.
-    """
-
-    # TODO: move to morphodict
-
-    def post(self, request):
-        orth = request.POST.get("orth")
-
-        # Tried to set to an unsupported orthography
-        if orth not in ORTHOGRAPHY.available:
-            return HttpResponse(status=HTTPStatus.BAD_REQUEST)
-
-        response = HttpResponse(status=HTTPStatus.NO_CONTENT)
-        response.set_cookie("orth", orth)
-        return response
 
 
 def redirect_search(request, query_string: str):
